style: introduce checkstyle for project
