+++ src/main/java/org/arsonal/accounting/controller/HelloController.java
@@ -1,6 +1,7 @@
 package org.arsonal.accounting.controller;
 
 import org.arsonal.accounting.model.service.Greeting;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
