+++ src/main/java/org/arsonal/accounting/dao/UserInfoDao.java
@@ -3,6 +3,6 @@ package org.arsonal.accounting.dao;
 import org.arsonal.accounting.model.persistence.UserInfo;
 
 
-public interface UserInfoDAO {
+public interface UserInfoDao {
     UserInfo getUserInfoById(Long id);
 }
