+++ src/main/java/org/arsonal/accounting/exception/InvalidParameterException.java
@@ -3,6 +3,10 @@ package org.arsonal.accounting.exception;
 import org.springframework.http.HttpStatus;
 
 public class InvalidParameterException extends ServiceException {
+    /**
+     * constructor for invalidParameterException.
+     * @param message thrown message.
+     */
     public InvalidParameterException(String message) {
         super(message);
         this.setErrorCode("INVALID_PARAMETER");
