+++ src/main/java/org/arsonal/accounting/model/common/UserInfo.java
@@ -3,7 +3,7 @@ package org.arsonal.accounting.model.common;
 import lombok.Builder;
 import lombok.Data;
 
-// 给Service/manager内部流转使用，Service/manager层
+// 给Service/manager内部流转使用，即Service/manager层
 @Data
 @Builder
 public class UserInfo {
